优化播放逻辑

+++ TMusic/Search (搜索)/Controllers/TArtistViewController.m
@@ -55,7 +55,7 @@
 
 -(UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath
 {
-     NSString *cellID = [NSString stringWithFormat:@"%lu",indexPath.row];
+     NSString *cellID = [NSString stringWithFormat:@"%lu",(long)indexPath.row];
     UITableViewCell *cell = [tableView dequeueReusableCellWithIdentifier:cellID];
     if (cell == nil) {
         cell =[[UITableViewCell alloc]initWithStyle:UITableViewCellStyleSubtitle reuseIdentifier:cellID];
